Add necessary methods to ReservationService

+++ core-module/src/main/java/org/kindness/coremodule/domain/reservation/ReservationService.java
@@ -10,8 +10,11 @@ import org.kindness.module.persistence.dao.impl.JdbcTableDao;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.swing.plaf.PanelUI;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
+import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -21,7 +24,7 @@ public class ReservationService {
     private final JdbcTableDao tableDao;
 
     @Transactional
-    public void createReservation(Reservation res) throws IllegalStateException {
+    public Long createReservation(Reservation res) throws IllegalStateException {
         if (res.getReservationStart().isBefore(LocalDateTime.now())){
             throw new IllegalStateException("Cant reserve before now");
         }
@@ -36,19 +39,46 @@ public class ReservationService {
         Restaurant restaurant = restaurantDao.findById(table.getRestaurantId()).orElseThrow(() ->
                 new IllegalStateException("No such restaurant"));
 
-        if(restaurant.getOpenTime().isBefore(LocalTime.now())
-                && restaurant.getCloseTime().isAfter(LocalTime.now())){
-            throw new IllegalStateException("The restaurant is not open at that time");
+        LocalTime startTime = res.getReservationStart().toLocalTime();
+        LocalTime endTime = res.getReservationEnd().toLocalTime();
+
+        if (startTime.isBefore(restaurant.getOpenTime()) || endTime.isAfter(restaurant.getCloseTime())) {
+            throw new IllegalStateException("Restaurant is not open at that time");
         }
 
-        reservationDao.insert(res);
+        return reservationDao.insertAndGetID(res);
     }
 
     @Transactional
-    public void cancelReservation(Long reservationId){
-        var savedRes = reservationDao.findById(reservationId).orElseThrow(() ->
-                new IllegalStateException("No such reservation found"));
+    public void cancelReservation(Long reservationId, Long userId){
+        var reservation = reservationDao.findById(reservationId)
+                .orElseThrow(() -> new IllegalStateException("Reservation not found"));
+
+        if (!reservation.getUserId().equals(userId)) {
+            throw new IllegalStateException("You cant access that reservation");
+        }
+
         reservationDao.deleteById(reservationId);
+    }
+
+    @Transactional(readOnly = true)
+    public List<Reservation> getAllUserReservations(Long userId) {
+        return reservationDao.findAll(userId);
+    }
+
+    @Transactional(readOnly = true)
+    public List<Reservation> getAllReservations(){
+        return reservationDao.findAll();
+    }
+
+    @Transactional(readOnly = true)
+    public Optional<Reservation> getUserReservation(Long reservationId, Long userId){
+        return reservationDao.findById(reservationId, userId);
+    }
+
+    @Transactional(readOnly = true)
+    public Optional<Reservation> getReservation(Long reservationId){
+        return reservationDao.findById(reservationId);
     }
 
 }
